Feature: split cartReducer and added Cart Component
