+++ client/components/SingleProduct.js
@@ -1,9 +1,10 @@
 import React, {Component} from 'react';
 import {connect} from 'react-redux';
-import {getSingleProduct, addToCartThunk} from '../store/products';
+import {getSingleProduct} from '../store/products';
 import {Container, Col, Row, Button} from 'reactstrap';
 import {priceToDollar} from '../utilities/convertPriceToDollars';
 import {RiShoppingCartLine} from 'react-icons/ri';
+import {addToCartThunk} from '../store/cart';
 
 class SingleProduct extends Component {
   constructor() {
