+++ client/components/user-home.js
@@ -1,19 +1,20 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import {connect} from 'react-redux'
+import React from 'react';
+import PropTypes from 'prop-types';
+import {connect} from 'react-redux';
+import axios from 'axios';
 
 /**
  * COMPONENT
  */
 export const UserHome = props => {
-  const {email} = props
+  const {email} = props;
 
   return (
     <div>
       <h3>Welcome, {email}</h3>
     </div>
-  )
-}
+  );
+};
 
 /**
  * CONTAINER
@@ -21,14 +22,14 @@ export const UserHome = props => {
 const mapState = state => {
   return {
     email: state.user.email
-  }
-}
+  };
+};
 
-export default connect(mapState)(UserHome)
+export default connect(mapState)(UserHome);
 
 /**
  * PROP TYPES
  */
 UserHome.propTypes = {
   email: PropTypes.string
-}
+};
