+++ client/routes.js
@@ -7,7 +7,8 @@ import {
   Signup,
   UserHome,
   AllProducts,
-  SingleProduct
+  SingleProduct,
+  Cart
 } from './components';
 import {me} from './store';
 
@@ -29,6 +30,7 @@ class Routes extends Component {
         <Route path="/signup" component={Signup} />
         <Route exact path="/products" component={AllProducts} />
         <Route exact path="/products/:id" component={SingleProduct} />
+        <Route exact path="/cart" component={Cart} />
         {isLoggedIn && (
           <Switch>
             {/* Routes placed here are only available after logging in */}
