+++ client/store/products.js
@@ -3,9 +3,6 @@ import Axios from 'axios';
 //ACTION TYPES
 const ALL_PRODUCTS = 'ALL_PRODUCTS';
 const SINGLE_PRODUCT = 'SINGLE_PRODUCT';
-const ADD_QUANTITY = 'ADD_QUANTITY';
-const SUB_QUANTITY = 'SUB_QUANTITY';
-const ADD_TO_CART = 'ADD_TO_CART';
 
 //ACTION CREATORS
 export const allProducts = products => {
@@ -18,27 +15,6 @@ export const allProducts = products => {
 export const singleProduct = product => {
   return {
     type: SINGLE_PRODUCT,
-    product
-  };
-};
-
-export const addQuantity = product => {
-  return {
-    type: ADD_QUANTITY,
-    product
-  };
-};
-
-export const subQuantity = product => {
-  return {
-    type: SUB_QUANTITY,
-    product
-  };
-};
-
-export const addedToCart = product => {
-  return {
-    type: ADD_TO_CART,
     product
   };
 };
@@ -66,20 +42,9 @@ export const getSingleProduct = productId => {
   };
 };
 
-export const addToCartThunk = product => {
-  return dispatch => {
-    try {
-      dispatch(addedToCart(product));
-    } catch (error) {
-      console.log(error);
-    }
-  };
-};
-
 const initialState = {
   allProducts: [],
-  selectedProduct: {},
-  cart: []
+  selectedProduct: {}
 };
 
 export default function productsReducer(state = initialState, action) {
@@ -93,31 +58,6 @@ export default function productsReducer(state = initialState, action) {
       return {
         ...state,
         selectedProduct: action.product
-      };
-    case ADD_TO_CART:
-      return {
-        ...state,
-        cart: [...state.cart, {quantity: 1, product: action.product}]
-      };
-    case SUB_QUANTITY:
-      return {
-        ...state,
-        cart: state.cart.map(product => {
-          if (product.product.name === action.product.name) {
-            return {...product, quantity: product.quantity - 1};
-          }
-          return product;
-        })
-      };
-    case ADD_QUANTITY:
-      return {
-        ...state,
-        cart: state.cart.map(product => {
-          if (product.product.name === action.product.name) {
-            return {...product, quantity: product.quantity + 1};
-          }
-          return product;
-        })
       };
     default:
       return state;
